dep: rm tags functionality
